storage update

+++ src/main/java/team/workflow/services/service/Impl/StorageServiceImpl.java
@@ -68,7 +68,6 @@ public class StorageServiceImpl implements StorageService {
                                             return Mono.just(0);
                                         }
                                     });
-
                                 }
                                 return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
                             }else {
@@ -78,15 +77,61 @@ public class StorageServiceImpl implements StorageService {
                         }
 
                 );
-        //return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
     }
 
     //库存解锁
+    //判定库存锁定的oid为自己的oid
     @Override
     public Mono<ResponseEntity> StorageUnlock(String jsonStr) {
         JSONObject object = JSON.parseObject(jsonStr);
         String pid= (String) object.get("pid");
-        Mono<Integer> res=productRepository.deleteOid(pid);
-        return Mono.just(new ResponseEntity(res,HttpStatus.OK));
+        String oid= (String) object.get("oid");
+        Mono<Product> product=productRepository.findById(pid);
+        //检测是否解锁成功
+        return product
+                .flatMap(product1 -> Mono.just(Optional.of(product1)))
+                .defaultIfEmpty(Optional.empty())
+                .flatMap(product1 -> {
+                       if(product1.isEmpty()){
+                           return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
+                        }
+                        List<Product> productList=product1.map(Collections::singletonList).orElse(Collections.emptyList());
+                       //检测是否为自己的oid
+                        if(productList.get(0).getOid().equals(oid)){
+                            //解锁，将oid设为null
+                            Mono<Integer> res=productRepository.deleteOid(pid);
+                            return  Mono.just(new ResponseEntity(res,HttpStatus.OK));
+
+                        }else {
+                            return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
+                        }
+                });
+    }
+
+    @Override
+    public Mono<ResponseEntity> StorageUpdate(String jsonStr) {
+        JSONObject object = JSON.parseObject(jsonStr);
+        String pid= (String) object.get("pid");
+        String oid= (String) object.get("oid");
+        Mono<Product> product=productRepository.findById(pid);
+        //检测是否解锁成功
+        return product
+                .flatMap(product1 -> Mono.just(Optional.of(product1)))
+                .defaultIfEmpty(Optional.empty())
+                .flatMap(product1 -> {
+                    if(product1.isEmpty()){
+                        return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
+                    }
+                    List<Product> productList=product1.map(Collections::singletonList).orElse(Collections.emptyList());
+                    //检测是否为自己的oid
+                    if(productList.get(0).getOid().equals(oid)){
+                        //库存-1
+                        Mono<Integer> res=productRepository.updateStorage(pid);
+                        return  Mono.just(new ResponseEntity(res,HttpStatus.OK));
+
+                    }else {
+                        return Mono.just(new ResponseEntity(HttpStatus.NOT_ACCEPTABLE));
+                    }
+                });
     }
 }
